Added ArmyPrint class
