+++ ArmyBuilder/Gui/ArmyView.xaml.cs
@@ -5,9 +5,6 @@ using Microsoft.Extensions.DependencyInjection;
 using ArmyBuilder.ViewModels;
 using ArmyBuilder.Domain;
 using ArmyBuilder.Dao;
-using QuestPDF.Fluent;
-using QuestPDF.Helpers;
-using QuestPDF.Infrastructure;
 
 namespace ArmyBuilder
 {
@@ -26,75 +23,11 @@ namespace ArmyBuilder
 
         private void Print_Click(object sender, RoutedEventArgs e)
         {
-            // Define the output file path
-            string outputPath = "example.pdf";
-            QuestPDF.Settings.License = LicenseType.Community;
-
-            // Create a PDF document
-            Document.Create(container =>
-            {
-                container.Page(page =>
-                {
-                    page.Size(PageSizes.A4);
-                    page.Margin(2, QuestPDF.Infrastructure.Unit.Centimetre);
-                    page.PageColor(Colors.White);
-                    page.DefaultTextStyle(x => x.FontSize(20));
-
-                    page.Header()
-                        .Text("Hello, World!")
-                        .SemiBold().FontSize(36).FontColor(Colors.Blue.Medium);
-
-                    page.Content()
-                        .Table(table =>
-                        {
-                            // Define columns
-                            table.ColumnsDefinition(columns =>
-                            {
-                                columns.RelativeColumn();
-                                columns.RelativeColumn();
-                                columns.RelativeColumn();
-                            });
-
-                            // Add header row
-                            table.Header(header =>
-                            {
-                                header.Cell().Element(CellStyle).Text("Header 1");
-                                header.Cell().Element(CellStyle).Text("Header 2");
-                                header.Cell().Element(CellStyle).Text("Header 3");
-
-                                static IContainer CellStyle(IContainer container)
-                                {
-                                    return container.DefaultTextStyle(x => x.SemiBold()).Padding(5).Background(Colors.Grey.Lighten3);
-                                }
-                            });
-
-                            // Add data rows
-                            for (int i = 0; i < 9; i++)
-                            {
-                                table.Cell().Element(CellStyle).Text($"Cell {i + 1}");
-                            }
-
-                            static IContainer CellStyle(IContainer container)
-                            {
-                                return container.Border(1).BorderColor(Colors.Grey.Lighten2).Padding(5);
-                            }
-                        });
-
-                    page.Footer()
-                        .AlignCenter()
-                        .Text(x =>
-                        {
-                            x.Span("Page ");
-                            x.CurrentPageNumber();
-                        });
-                });
-            })
-            .GeneratePdf(outputPath);
-
-            // Inform the user
-            Console.WriteLine("PDF created successfully!");
+            var armyViewModel = DataContext as ArmyViewModel;
+            var army = armyViewModel.ArmyTreeViewModel.Army;
+            var armyPrint = new ArmyPrint();
+            armyPrint.PrintArmy(army);
         }
-
         private void QuitEdition_Click(object sender, RoutedEventArgs e)
         {
             var armyViewModel = DataContext as ArmyViewModel;
@@ -140,7 +73,7 @@ namespace ArmyBuilder
                     ArmyViewModel armyViewModel = DataContext as ArmyViewModel;
                     ArmyTreeViewModel armyTreeViewModel = armyViewModel.ArmyTreeViewModel;
                     Army army = armyTreeViewModel.Army;
-                    ArmyBuilder.Domain.Unit unit = armyViewModel.CreateUnit(army, clonedMainModel);
+                    Unit unit = armyViewModel.CreateUnit(army, clonedMainModel);
                     ArmyTreeNode armyTreeNode = armyTreeViewModel.Root[0];
                     armyTreeNode.AddUnit(unit);
                 }
@@ -162,7 +95,7 @@ namespace ArmyBuilder
                         MainModel clonedMainModel = droppedMainModel.Clone();
                         Army army = armyTreeNode.Army;
                         var armyViewModel = DataContext as ArmyViewModel;
-                        ArmyBuilder.Domain.Unit unit = armyViewModel.CreateUnit(army, clonedMainModel);
+                        Unit unit = armyViewModel.CreateUnit(army, clonedMainModel);
                         armyTreeNode.AddUnit(unit);
                     }
                 }
@@ -181,7 +114,7 @@ namespace ArmyBuilder
                     if (unitTreeNode != null)
                     {
                         MainModel clonedMainModel = droppedMainModel.Clone();
-                        ArmyBuilder.Domain.Unit unit = unitTreeNode.Unit;
+                        Unit unit = unitTreeNode.Unit;
                         var armyViewModel = DataContext as ArmyViewModel;
                         armyViewModel.AddMainModel(unit.Id, clonedMainModel);
                         unitTreeNode.AddMainModel(clonedMainModel);
